+++ libsrc/math/sin.c
@@ -21,7 +21,7 @@ double sin(double x) {
 
   double v = 0;
   double xx = x * x;
-  for (int i = 0; i < sizeof(TABLE) / sizeof(*TABLE); ++i) {
+  for (int i = 0; i < (int)(sizeof(TABLE) / sizeof(*TABLE)); ++i) {
     v += x * TABLE[i];
     x *= xx;
   }
