+++ libsrc/stdio/fputc.c
@@ -4,7 +4,7 @@
 
 int fputc(int c, FILE* fp) {
   fp->wbuf[fp->wp++] = c;
-  if (fp->wp >= sizeof(fp->wbuf) ||
+  if (fp->wp >= (int)sizeof(fp->wbuf) ||
       (c == '\n' && !(fp->flag & FF_BINARY))) {
     if (fflush(fp) == EOF)
       return EOF;
